requirements.txt
bottle
gunicorn
PyYAML
requests
urllib3